+++ erpnext/manufacturing/doctype/bom/bom.js
@@ -132,24 +132,13 @@ frappe.ui.form.on("BOM", {
 
 	make_work_order: function(frm) {
 		let fields = [];
-		if (frm.doc.manufacturing_type == "Process") {
-			fields = [{
-				fieldtype: 'Float',
-				label: __('Raw Material Qty'),
-				fieldname: 'qty',
-				reqd: 1,
-				default: 1
-			}];
-		}
-		else {
-			fields = [{
-				fieldtype: 'Float',
-				label: __('Qty To Manufacture'),
-				fieldname: 'qty',
-				reqd: 1,
-				default: 1
-			}];
-		}
+		fields = [{
+			fieldtype: 'Float',
+			label: frm.doc.manufacturing_type == "Process" ? __('Raw Material Qty') : __('Qty To Manufacture'),
+			fieldname: 'qty',
+			reqd: 1,
+			default: 1
+		}];
 
 		frappe.prompt(fields, data => {
 			frappe.call({
@@ -157,10 +146,10 @@ frappe.ui.form.on("BOM", {
 				args: {
 					bom_no: frm.doc.name,
 					item: frm.doc.item,
-					manufacturing_type: frm.doc.manufacturing_type,
-					finished_good_qty: frm.doc.items[0].qty || 0,
 					qty: data.qty || 0.0,
 					project: frm.doc.project,
+					manufacturing_type: frm.doc.manufacturing_type,
+					finished_goods_qty: frm.doc.items[0].qty || 0,
 				},
 				freeze: true,
 				callback: function(r) {
