--- erpnext/patches/v13_0/set_manufacturing_type_in_boms.py
@@ -1,11 +1,0 @@
-import frappe
-
-def execute():
-	frappe.db.sql("""
-	UPDATE
-		`tabBOM`
-	SET
-		manufacturing_type = "Discrete"
-	WHERE
-		manufacturing_type IS NULL
-""")
