--- erpnext/patches/v13_0/set_manufacturing_type_in_stock_entries.py
@@ -1,11 +1,0 @@
-import frappe
-
-def execute():
-	frappe.db.sql("""
-	UPDATE
-		`tabStock Entry`
-	SET
-		manufacturing_type = "Discrete"
-	WHERE
-		manufacturing_type IS NULL
-""")
