--- erpnext/patches/v13_0/set_manufacturing_type_in_work_orders.py
@@ -1,11 +1,0 @@
-import frappe
-
-def execute():
-	frappe.db.sql("""
-	UPDATE
-		`tabWork Order`
-	SET
-		manufacturing_type = "Discrete"
-	WHERE
-		manufacturing_type IS NULL
-""")
